Create add-remove.js
