+++ src/index.html
@@ -35,6 +35,6 @@
     </form>
     <a href="#" class="clear">Clear all completed</a>
   </div>
-  <script src="index.js"></script>
+  <script src="index.js" type="module"></script>
 </body>
 </html>
